move sight_list_Screen to ui/screen
