+++ lib/ui/screen/sight_list_screen.dart
@@ -1,0 +1,43 @@
+import 'package:flutter/material.dart';
+import 'package:places/mocks.dart';
+import 'package:places/ui/screen/sight_card.dart';
+
+class SightListScreen extends StatefulWidget {
+  SightListScreen({Key key}) : super(key: key);
+
+  @override
+  _SightListScreenState createState() => _SightListScreenState();
+}
+
+class _SightListScreenState extends State<SightListScreen> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        backgroundColor: Colors.transparent,
+        elevation: 0,
+        toolbarHeight: 128,
+        centerTitle: false,
+        title: const Text(
+          'Список\nинтересных мест',
+          style: TextStyle(
+            color: Color(0xFF3B3E5B),
+            fontSize: 32,
+            fontWeight: FontWeight.bold,
+          ),
+          maxLines: 2,
+          overflow: TextOverflow.ellipsis,
+        ),
+      ),
+      body: SingleChildScrollView(
+        child: Column(
+          children: [
+            SightCard(sight: mocks[0]),
+            SightCard(sight: mocks[1]),
+            SightCard(sight: mocks[2]),
+          ],
+        ),
+      ),
+    );
+  }
+}
